DBStringLen: Return string at max length x - subtle bug in here
When inserting a string into a database column of type character varying,
we have to pay attention not to exceed the length limit. However simply
treating the string simply as a byte array and slice cheaply will yield
errors with databases which check for valid character encoding when split
at an invalid rune boundary.